Factor out common logic from IntegerArray into an ArrayIndex superclass
This will be the common superclass of IntegerArray and BooleanArray.
