+++ ndindex/integerarray.py
@@ -1,11 +1,11 @@
 import warnings
 
-from numpy import (ndarray, asarray, integer, bool_, intp, array2string,
-                   empty, zeros)
+from numpy import ndarray, asarray, integer, bool_, intp, empty, zeros
 
-from .ndindex import NDIndex, asshape
+from .array import ArrayIndex
+from .ndindex import asshape
 
-class IntegerArray(NDIndex):
+class IntegerArray(ArrayIndex):
     """
     Represents an integer array.
 
@@ -32,58 +32,7 @@ class IntegerArray(NDIndex):
            [1, 2]])
 
     """
-    def _typecheck(self, idx, shape=None):
-        if shape is not None:
-            if idx != []:
-                raise ValueError("The shape argument is only allowed for empty arrays (idx=[])")
-            shape = asshape(shape)
-            if 0 not in shape:
-                raise ValueError("The shape argument must be an empty shape")
-            idx = empty(shape, dtype=intp)
-
-        if isinstance(idx, (list, ndarray, bool, integer, int, bool_)):
-            # Ignore deprecation warnings for things like [1, []]. These will be
-            # filtered out anyway since they produce object arrays.
-            with warnings.catch_warnings(record=True):
-                a = asarray(idx)
-                if a is idx:
-                    a = a.copy()
-                if isinstance(idx, list) and 0 in a.shape:
-                    a = a.astype(intp)
-            if issubclass(a.dtype.type, integer):
-                if a.dtype != intp:
-                    a = a.astype(intp)
-                a.flags.writeable = False
-                return (a,)
-            if a.dtype == bool_:
-                raise TypeError("Boolean array passed to IntegerArray. Use BooleanArray instead.")
-            raise TypeError("The input array must have an integer dtype.")
-        raise TypeError("IntegerArray must be created with an array of integers")
-
-    @property
-    def raw(self):
-        return self.args[0]
-
-    @property
-    def array(self):
-        """
-        Return the NumPy array of self.
-
-        This is the same as `self.args[0]`.
-        """
-        return self.args[0]
-
-    @property
-    def shape(self):
-        """
-        Return the shape of the array of self.
-
-        This is the same as self.array.shape. Note that this is **not** the
-        same as the shape of an array that is indexed by self. Use
-        :meth:`newshape` to get that.
-
-        """
-        return self.array.shape
+    dtype = intp
 
     def reduce(self, shape=None, axis=0):
         """
@@ -159,21 +108,3 @@ class IntegerArray(NDIndex):
                 return True
 
         return 0 in self.shape
-
-    # The repr form recreates the object. The str form gives the truncated
-    # array string and is explicitly non-valid Python (doesn't have commas).
-    def __repr__(self):
-        if 0 not in self.shape:
-            arg = repr(self.array.tolist())
-        else:
-            arg = f"[], shape={self.shape}"
-        return f"{self.__class__.__name__}({arg})"
-
-    def __str__(self):
-        return (self.__class__.__name__
-                + "("
-                + array2string(self.array).replace('\n', '')
-                + ")")
-
-    def __hash__(self):
-        return hash(self.array.tobytes())
